Completed part 10 now, CPU is practically fully implemented now. Have to fix some debugging, but after that it is ready to go :D

+++ lib/stack.c
@@ -7,7 +7,7 @@ void stack_push (u8 data) {
     bus_write(cpu_get_regs()->sp, data);
 }
 
-void stack_push16 (u8 data) {
+void stack_push16 (u16 data) {
     stack_push((data >> 8) & 0xFF);
     stack_push(data & 0xFF);
 }
@@ -16,7 +16,7 @@ u8 stack_pop () {
     return bus_read(cpu_get_regs()->sp++);
 }
 
-u8 stack_pop16 () {
+u16 stack_pop16 () {
     u16 lo = stack_pop();
     u16 hi = stack_pop();
     return (hi << 8) | lo;
